Refactor resend verification button and middleware
