+++ app/Http/Controllers/HomeController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Alumnus;
+use App\User;
 use Illuminate\Http\Request;
 use App\Post;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
+use Jrean\UserVerification\Facades\UserVerification;
 
 class HomeController extends Controller
 {
@@ -41,5 +43,20 @@ class HomeController extends Controller
             else
                 return view('home')->withPosts($posts);
         }
+    }
+
+    /**
+     * Resend verification token
+     *
+     * @param $id user id
+     * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
+     */
+    public function resendVerification($id)
+    {
+        $user = User::find($id);
+        UserVerification::generate($user);
+        UserVerification::send($user, 'Please verify to complete registration at the DePaul Alumni Outreach System.');
+        Session::flash('message', 'You will receive your verification email shortly.');
+        return view('auth.errors.not-verified');
     }
 }
